test(flags): merge tests into one file
