--- flags/stop_early_test.ts
@@ -1,15 +1,0 @@
-// Copyright 2018-2021 the Deno authors. All rights reserved. MIT license.
-import { assertEquals } from "../testing/asserts.ts";
-import { parse } from "./mod.ts";
-
-// stops parsing on the first non-option when stopEarly is set
-Deno.test("stopParsing", function (): void {
-  const argv = parse(["--aaa", "bbb", "ccc", "--ddd"], {
-    stopEarly: true,
-  });
-
-  assertEquals(argv, {
-    aaa: "bbb",
-    _: ["ccc", "--ddd"],
-  });
-});
